refactor: remove Google Analytics, keep backend analytics only
- Replace GoogleAnalyticsContext with AnalyticsContext: drop react-ga4
  initialization, GA events, and pageview sending; events now go only
  to the backend analytics endpoint
- Rename useGoogleAnalytics -> useAnalytics and sendGaEvent ->
  sendAnalyticsEvent across all consumers
- Remove unused sendPageViewEvent (GA-only, no callers)
- Remove GA_MEASUREMENT_ID from config and measurementId prop from Layout
- Uninstall react-ga4 dependency
- Type analytics events properly; allow extra fields (e.g. hasResults)
  to keep backend payloads unchanged
