+++ package.json
@@ -24,7 +24,6 @@
     "react": "^18.3.1",
     "react-device-detect": "^2.2.3",
     "react-dom": "^18.3.1",
-    "react-ga4": "^2.1.0",
     "react-i18next": "^15.0.1",
     "react-loader-spinner": "^6.1.6",
     "react-markdown": "8.0.6",
