+++ src/hooks/useAnalytics.tsx
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { AnalyticsContext, AnalyticsContextType } from "../context/AnalyticsContext";
+
+export const useAnalytics = (): AnalyticsContextType => {
+    const context = useContext(AnalyticsContext);
+    if (context === undefined) {
+        throw new Error('useAnalytics must be used within an AnalyticsProvider');
+    }
+    return context;
+};
