+++ src/types/types.d.ts
@@ -115,11 +115,6 @@ interface Movements {
     deals: Deal[]
 }
 
-interface GoogleAnalyticsContextType {
-    sendGaEvent: (name: EventName, event: GAEvent) => void;
-    sendPageViewEvent: (path: string) => void;
-}
-
 interface Link {
     href: string,
     linkText: string
